@@ -27,13 +27,11 @@
         "primereact": "7.1.0",
         "prismjs": "^1.27.0",
         "quagga": "^0.12.1",
-        "react": "^17.0.1",
+        "react": "18",
         "react-app-polyfill": "^1.0.6",
         "react-dnd": "^15.1.1",
         "react-dnd-html5-backend": "^15.1.2",
-        "react-dom": "^17.0.1",
-        "react-qr-reader": "^3.0.0-beta-1",
-        "react-qr-scanner": "^1.0.0-alpha.11",
+        "react-dom": "18",
         "react-router-dom": "^5.2.0",
         "react-scripts": "^5.0.0",
         "react-transition-group": "^4.4.1",
