@@ -1,18 +1,18 @@
-import 'react-app-polyfill/ie11';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-//import * as serviceWorker from './serviceWorker';
-import { HashRouter } from 'react-router-dom'
-import ScrollToTop from './ScrollToTop';
+import React from "react";
+import { createRoot } from "react-dom/client";
+import { HashRouter } from "react-router-dom";
+import App from "./App";
+import ScrollToTop from "./ScrollToTop";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
     <HashRouter>
         <ScrollToTop>
-            <App></App>
+            <App />
         </ScrollToTop>
-    </HashRouter>,
-    document.getElementById('root')
+    </HashRouter>
 );
 
 // If you want your app to work offline and load faster, you can change
