@@ -4,11 +4,8 @@ import TextInput from "./TextInput";
 import { TabPanel, TabView } from "primereact/tabview";
 import { Dialog } from "primereact/dialog";
 import { Toast } from "primereact/toast";
-import { Toolbar } from "primereact/toolbar";
 import InputTextArea from "./InputTextArea";
-import { FileUpload } from "primereact/fileupload";
 import PoliticalPartyService from "../service/PoliticalPartyService";
-import imageToBase64 from "image-to-base64/browser";
 import { Image } from "primereact/image";
 
 export default function AddPoliticalParty({ setPoliticalParties, buttonName = "Save", buttonIcon = "pi pi-save", show = false, setShow }) {
@@ -169,16 +166,16 @@ export default function AddPoliticalParty({ setPoliticalParties, buttonName = "S
                                 <div className="grid">
                                     <div className="col-12  lg:col-4">
                                         <label htmlFor="description">Logo/Symbol</label> <br></br>
-                                        <div style={{ visibility: "hidden" }}>LOGO</div>  
+                                        <div style={{ visibility: "hidden" }}>LOGO</div>
                                         <React.Fragment>
                                             <Button label={form.FileName.trim().length === 0 ? "Select a file" : form.FileName} onClick={onBtnClick} className="p-button-success" icon={form.FileName.trim().length === 0 ? "pi pi-plus" : ""} />
                                             <input ref={inputFileRef} type={"file"} onChange={(e) => onUploadHandler(e.target)} style={{ display: "none" }}></input>
                                         </React.Fragment>
-                                        <div style={{ visibility: "hidden" }}>LOGO</div>  
+                                        <div style={{ visibility: "hidden" }}>LOGO</div>
                                         {form.FileName.trim().length === 0 ? "" : <Image preview={true} src={`data:image/png;base64,${form.Logo}`} alt="Logo" width="100px" style={{ width: "100px", objectFit: "cover" }} />}
                                     </div>
                                 </div>
-                              
+
                             </TabPanel>
                         </TabView>
                     </div>
